refactor: migrate to component-based GUI architecture
