+++ src/gui/forwards.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 
 namespace gui {
     enum class GuiPage {
