+++ src/gui/gui.h
@@ -11,17 +11,22 @@
 
 #include "gui/forwards.h"
 #include "gui/constants.h"
+#include "gui/utils.h"
 #include "gui/sag_analysis_service.h"
+#include "gui/sidebar_renderer.h"
+#include "gui/content_router.h"
+#include "gui/menu_bar_controller.h"
+#include "gui/sag_analysis_controller.h"
+#include "gui/graphics_backend.h"
+#include "gui/debug_log_viewer.h"
+#include "gui/app_info_dialog.h"
 #include "dde/client.h"
+#include "dde/dde_connection_manager.h"
 
 class Logger;
 
 namespace gui {
-    // Free utility functions
-    const char* getUnitString(int unitCode, bool full = false);
-    const char* getRayAimingTypeString(int rayAimingType);
-    void HelpMarker(const char* desc);
-    void renderPageHeader(GuiPage currentPage);
+    // Free utility functions (declared in utils.h, included above)
 
     class GuiManager {
         public:
@@ -32,9 +37,8 @@ namespace gui {
             void render();
             void updateDpiStyle(float dpiScale);
 
-            void renderNavbar();
+            // Delegated to components
             void renderSidebar();
-            void renderDDEStatusFrame();
             void renderContent();
             void renderDebugLog();
             void setPopupPosition();
@@ -52,10 +56,21 @@ namespace gui {
             HWND m_hwndClient;                                    // DDE client window handle
             ZemaxDDE::ZemaxDDEClient* m_zemaxDDEClient;           // Pointer to a DDE client instance
             Logger& m_logger;                                     // Logger instance (dependency injection)
+
+            // Components (New approach)
+            GraphicsBackend m_graphics;
             std::unique_ptr<SagAnalysisService> m_sagService;     // Sag analysis service (owned by GuiManager)
+            std::unique_ptr<SagAnalysisController> m_sagController;
+            std::unique_ptr<DdeConnectionManager> m_ddeConnectionManager;
+            std::unique_ptr<MenuBarController> m_menuBarController;
+            std::unique_ptr<SidebarRenderer> m_sidebarRenderer;
+            std::unique_ptr<ContentRouter> m_contentRouter;
+            std::unique_ptr<DebugLogViewer> m_debugLogViewer;
+            std::unique_ptr<AppInfoDialog> m_appInfoDialog;
 
             GuiPage m_currentPage = GuiPage::OpticalSystemInfo;
 
+            // State
             bool m_showUpdatesPopup{false};                       // Display flag for popup 'Check for Updates'
             bool m_showAboutPopup{false};                         // Display flag for popup 'Check for About'
     };
